requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-crontab==0.7.1
django-crum==0.7.9
django-dbbackup==4.0.0b0
django-google-maps==0.13.0
django-grappelli==3.0.3
gunicorn==20.1.0
ply==3.11
pyasn1==0.4.8
pycryptodomex==3.14.1
pycurl==7.45.1
pysmi==0.3.4
pysnmp==4.4.12
python-dotenv==0.20.0
python-telegram==0.15.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
